refactor: optimize file directories
