+++ src/components/EmptyDataBox/styles.module.scss
