+++ src/components/Modal/SimpleTipModal/index.tsx
