+++ src/components/Modal/SimpleTipModal/styles.module.scss
