+++ src/components/Modal/SynchronizingChainModal/styles.module.scss
