+++ src/pageComponents/WithdrawContent/ContractAddress/index.tsx
@@ -2,7 +2,7 @@ import clsx from 'clsx';
 import styles from './styles.module.scss';
 import { CONTRACT_ADDRESS } from 'constants/deposit';
 import { QuestionMarkIcon } from 'assets/images';
-import ViewContractAddressModal from 'pageComponents/Modal/ViewContractAddressModal';
+import ViewContractAddressModal from 'components/Modal/ViewContractAddressModal';
 import { useCallback, useState } from 'react';
 import { formatStr2Ellipsis } from 'utils/format';
